@@ -1,4 +1,7 @@
 #!/usr/bin/python
+# Copyright 2025 The Levanter Authors
+# SPDX-License-Identifier: Apache-2.0
+
 import argparse
 import getpass
 import os
@@ -28,7 +31,7 @@ def main():
     cli.add_arg(
         parser, config, ["--autodelete"], default=False, action="store_true", help="Delete TPU after job completes."
     )
-    cli.add_arg(parser, config, ["--docker_base_image"], default="ghcr.io/stanford-crfm/levanter-base:latest")
+    cli.add_arg(parser, config, ["--docker_base_image"], default="ghcr.io/marin-community/levanter-base:latest")
     cli.add_arg(parser, config, ["--docker_repository"], default="levanter")
     cli.add_arg(parser, config, ["--foreground"], default=False, action="store_true")
     cli.add_arg(parser, config, ["--image_name"], default=f"levanter-{getpass.getuser()}")
