@@ -1,3 +1,6 @@
+# Copyright 2025 The Levanter Authors
+# SPDX-License-Identifier: Apache-2.0
+
 """
 This module contains code for running the [EleutherAI LM Evaluation Harness](https://github.com/EleutherAI/lm-evaluation-harness)
 inside Levanter runs. The EleutherAI LM Evaluation Harness is a tool for evaluating language models on a variety of tasks.
@@ -24,17 +27,17 @@ import tempfile
 import typing
 from dataclasses import dataclass
 from functools import cached_property
-from typing import Iterator, List, Optional, Tuple
+from typing import Iterator, List, Optional, Tuple, Union
 
 import equinox as eqx
+import haliax
 import jax
 import jax.numpy as jnp
+import jax.random as jrandom
 import jmp
 import numpy as np
-from jax.sharding import PartitionSpec
-
-import haliax
 from haliax import NamedArray
+from jax.sharding import PartitionSpec
 
 import levanter.tracker
 from levanter.compat.hf_checkpoints import HFCheckpointConverter, load_tokenizer
@@ -44,26 +47,31 @@ from levanter.data.packing import (
     per_segment_correct,
     per_segment_loss,
 )
+from levanter.inference.engine import InferenceEngine, InferenceEngineConfig
+from levanter.inference.engine import Request as GenRequest
+from levanter.inference.jit_scheduler import SeqDecodingParams
+from levanter.inference.utils import INVALID
 from levanter.models.gpt2 import Gpt2Config
 from levanter.models.loss import next_token_loss
 from levanter.utils.background_iterable import BackgroundIterator
 from levanter.utils.hf_utils import HfTokenizer
 from levanter.utils.py_utils import set_global_rng_seeds
 
-
 try:
     from lm_eval import evaluator
     from lm_eval.api.instance import Instance
-    from lm_eval.api.model import LM
+    from lm_eval.api.model import TemplateLM
+    from lm_eval.models.utils import handle_stop_sequences, postprocess_generated_text
 except ImportError:
-    LM = object
+    TemplateLM = object
     Instance = object
     evaluator = object
-
-from tqdm_loggable.auto import tqdm
+    handle_stop_sequences = None
+    postprocess_generated_text = None
 
 import haliax as hax
 from haliax.partitioning import ResourceMapping, round_axis_for_partitioning
+from tqdm_loggable.auto import tqdm
 
 import levanter.config
 from levanter.callbacks import StepInfo
@@ -75,8 +83,52 @@ from levanter.trainer import TrainerConfig
 from levanter.utils.jax_utils import broadcast_shard, use_cpu_device
 from levanter.utils.tree_utils import inference_mode
 
-
 logger = logging.getLogger(__name__)
+
+
+@dataclass(frozen=True)
+class SampleLoggingConfig:
+    """Configuration for controlling how many sample generations we keep per benchmark."""
+
+    log_all: bool = False
+    """If True, keep every sample that the harness sees."""
+    max_samples_per_benchmark: int | None = None
+    """Optional cap on the number of samples to store for each benchmark."""
+
+    def __post_init__(self):
+        if self.max_samples_per_benchmark is not None and self.max_samples_per_benchmark < 0:
+            raise ValueError("max_samples_per_benchmark must be non-negative or None")
+
+    def should_log(self) -> bool:
+        """Return True if any sample logging should occur."""
+        if self.log_all:
+            return True
+        return self.max_samples_per_benchmark is not None and self.max_samples_per_benchmark > 0
+
+    def allow_more(self, current_count: int) -> bool:
+        """Return True if we should store another sample for the benchmark."""
+        if not self.should_log():
+            return False
+        if self.log_all:
+            return True
+        assert self.max_samples_per_benchmark is not None
+        return current_count < self.max_samples_per_benchmark
+
+
+@dataclass(frozen=True)
+class ProfilerConfig:
+    """Configuration for JAX profiler during evaluation."""
+
+    enabled: bool = False
+    """If True, enable profiling during evaluation."""
+    start_step: int = 0
+    """Step at which to start profiling."""
+    num_steps: int = 0
+    """Number of steps to profile."""
+    profile_path: str = "/tmp/levanter_profile"
+    """Path to save profiler traces."""
+    perfetto_link: bool = False
+    """If True, create a Perfetto link for the trace."""
 
 
 # OK, so LM-Eval-Harness is not deterministic. This means we can't just run it on different workers and expect the
@@ -98,7 +150,17 @@ class _LmEvalHarnessWorker:
     """
 
     def __init__(
-        self, EvalBatch, EvalPos, model, axis_resources, tokenizer, mp, max_packed_segments, apply_chat_template=False
+        self,
+        EvalBatch,
+        EvalPos,
+        model,
+        axis_resources,
+        tokenizer,
+        mp,
+        max_packed_segments,
+        generation_kwargs=None,
+        sample_logging_config: SampleLoggingConfig | None = None,
+        profiler_config: ProfilerConfig | None = None,
     ):
         self.tokenizer = tokenizer
         self.max_packed_segments = max_packed_segments
@@ -108,7 +170,9 @@ class _LmEvalHarnessWorker:
         self.axis_resources = axis_resources
         self.mp = mp
         self.max_packed_segments = max_packed_segments
-        self.apply_chat_template = apply_chat_template
+        self._generation_kwargs = generation_kwargs or {"max_gen_toks": 256, "temperature": 0.0, "n": 1, "seed": None}
+        self.sample_logging_config = sample_logging_config or SampleLoggingConfig()
+        self.profiler_config = profiler_config or ProfilerConfig()
 
         self._dummy_batch = _make_dummy_batch(EvalBatch, EvalPos)
 
@@ -166,7 +230,12 @@ class _LmEvalHarnessWorker:
             _eval_loglikelihood, axis_resources=axis_resources, out_axis_resources={}
         )
 
-    def make_harness_lm(self, apply_chat_template: bool = False):
+    @property
+    def max_gen_toks(self) -> int:
+        """Backward compatibility property for max_gen_toks."""
+        return self._generation_kwargs.get("max_gen_toks", 256)
+
+    def make_harness_lm(self):
         if jax.process_index() == 0:
             return LevanterHarnessLM(self)
         else:
@@ -227,7 +296,7 @@ class _Message:
 
 
 def _get_segments_this_batch(batch, max_segments_per_ex):
-    unique_segs = np.unique(batch.attn_mask.segment_ids.array).tolist()
+    unique_segs = np.unique(batch.attn_mask.segment_ids[0].array).tolist()
     # + 1 because we use -1 as a padding value for segments and allow that
     if len(unique_segs) > max_segments_per_ex + 1:
         raise ValueError(f"Too many segments in batch: {len(unique_segs)}")
@@ -244,14 +313,153 @@ def _get_padding_count(batch, pad_token_id):
     return padding_count, total_tokens
 
 
-class LevanterHarnessLM(LM):
+class LevanterHarnessLM(TemplateLM):
     def __init__(self, leader: _LmEvalHarnessWorker):
         super().__init__()
         self.leader = leader
+        # Storage for prompts and generations to include in outputs
+        self.sample_outputs: dict[str, list[dict]] = {}
+        self.sample_logging_config = leader.sample_logging_config
+        self.profiler_config = leader.profiler_config
+        self._current_step = 0
+        self._profiler_started = False
 
     tokenizer = property(lambda self: self.leader.tokenizer)
     EvalBatch = property(lambda self: self.leader.EvalBatch)
     EvalPos = property(lambda self: self.leader.EvalPos)
+
+    @property
+    def tokenizer_name(self) -> str:
+        """Return a string identifier for the tokenizer/chat template."""
+        if hasattr(self.tokenizer, "name_or_path"):
+            return self.tokenizer.name_or_path
+        elif hasattr(self.tokenizer, "model_name"):
+            return self.tokenizer.model_name
+        else:
+            return "unknown_tokenizer"
+
+    def chat_template(self, chat_template: str | None = None) -> str | None:
+        """
+        Return the chat template for this model.
+
+        Args:
+            chat_template: Optional override for chat template
+
+        Returns:
+            The chat template string, or None if not available
+        """
+        if chat_template is not None:
+            return chat_template
+
+        # Try to get the chat template from the tokenizer
+        if hasattr(self.tokenizer, "chat_template") and self.tokenizer.chat_template is not None:
+            return self.tokenizer.chat_template
+
+        # If no chat template is available, return None
+        return None
+
+    @property
+    def max_gen_toks(self):
+        return self.leader.max_gen_toks
+
+    @property
+    def generation_kwargs(self):
+        """Get the generation kwargs from the worker."""
+        return self.leader._generation_kwargs
+
+    def apply_chat_template(self, chat_history: list[dict], **kwargs) -> str:
+        """
+        Apply chat template to format a conversation history.
+
+        Args:
+            chat_history: List of messages in the format [{"role": "user", "content": "..."}, ...]
+            **kwargs: Additional arguments to pass to the tokenizer's apply_chat_template method
+
+        Returns:
+            Formatted string ready for tokenization
+        """
+        return self.tokenizer.apply_chat_template(
+            chat_history,
+            tokenize=False,
+            add_generation_prompt=kwargs.get("add_generation_prompt", True),
+            **{k: v for k, v in kwargs.items() if k != "add_generation_prompt"},
+        )
+
+    @property
+    def eot_token_id(self) -> int:
+        """Return the end-of-text token ID."""
+        return self.tokenizer.eos_token_id
+
+    def set_current_task(self, task_name: str):
+        """Set the current task name for organizing sample outputs."""
+        self._current_task = task_name
+        if self.sample_logging_config.should_log() and task_name not in self.sample_outputs:
+            self.sample_outputs[task_name] = []
+
+    def get_sample_outputs(self) -> dict[str, list[dict]]:
+        """Get all stored sample outputs."""
+        return self.sample_outputs
+
+    def clear_sample_outputs(self):
+        """Clear all stored sample outputs."""
+        self.sample_outputs.clear()
+
+    def _prepare_bucket(self, task_name: str) -> list[dict[str, str]] | None:
+        if not self.sample_logging_config.should_log():
+            return None
+
+        bucket = self.sample_outputs.setdefault(task_name, [])
+        if self.sample_logging_config.allow_more(len(bucket)):
+            return bucket
+
+        return None
+
+    def _log_profiler_artifact(self):
+        """Log profiler artifact to the tracker."""
+        levanter.tracker.current_tracker().log_artifact(self.profiler_config.profile_path, type="jax_profile")
+
+    def _handle_profiler_step(self):
+        """Check if we should start or stop the profiler at this step."""
+        if not self.profiler_config.enabled:
+            return
+
+        start_step = self.profiler_config.start_step
+        num_steps = self.profiler_config.num_steps
+        end_step = start_step + num_steps
+
+        # Start profiler at start_step
+        if self._current_step == start_step and not self._profiler_started:
+            _create_perfetto_link = self.profiler_config.perfetto_link and jax.process_index() == 0
+
+            import os
+
+            os.makedirs(self.profiler_config.profile_path, exist_ok=True)
+
+            logger.info(f"Starting profiler at step {self._current_step} (will profile until step {end_step})")
+            jax.profiler.start_trace(
+                self.profiler_config.profile_path,
+                create_perfetto_link=_create_perfetto_link,
+                create_perfetto_trace=True,
+            )
+            self._profiler_started = True
+
+        # Stop profiler at end_step
+        elif self._current_step == end_step and self._profiler_started:
+            logger.info(f"Stopping profiler at step {self._current_step}")
+            jax.profiler.stop_trace()
+            self._profiler_started = False
+            self._log_profiler_artifact()
+
+    def _stop_profiler_if_needed(self):
+        """Ensure profiler is stopped if it was started."""
+        if self._profiler_started:
+            logger.info("Stopping profiler (end of evaluation).")
+            jax.profiler.stop_trace()
+            self._profiler_started = False
+            self._log_profiler_artifact()
+
+    def _loglikelihood_tokens(self, requests, disable_tqdm: bool = False):
+        raise NotImplementedError("_loglikelihood_tokens is not yet supported")
 
     def loglikelihood(self, requests: list[Instance]) -> list[tuple[float, bool]]:
         """
@@ -263,9 +471,21 @@ class LevanterHarnessLM(LM):
             logger.warning("No pad token set. Setting to eos token.")
             self.tokenizer.pad_token_id = self.tokenizer.eos_token_id
 
-        packed = _pack_requests(
-            requests, self.tokenizer, self.EvalPos, self.leader.max_packed_segments, self.leader.apply_chat_template
-        )
+        current_task = getattr(self, "_current_task", "loglikelihood_task")
+        for request in requests:
+            bucket = self._prepare_bucket(current_task)
+            if bucket is None:
+                break
+            prompt = request.args[0]
+            continuation = request.args[1]
+            bucket.append(
+                {
+                    "prompt": prompt,
+                    "generation": continuation,
+                }
+            )
+
+        packed = _pack_requests(requests, self.tokenizer, self.EvalPos, self.leader.max_packed_segments)
         packed_iterator = stack_batches(iter(packed), self.EvalPos, self.EvalBatch)
         packed_iterator = BackgroundIterator(packed_iterator, max_capacity=1024)
 
@@ -279,6 +499,9 @@ class LevanterHarnessLM(LM):
         total_tokens_seen = 0
         pbar = tqdm(total=total_tokens_expected, desc="loglikelihood", unit="tok")
         for q, batch in enumerate(packed_iterator):
+            # Handle profiler start/stop based on step
+            self._handle_profiler_step()
+
             segments_this_batch = _get_segments_this_batch(
                 batch, self.leader.max_packed_segments * self.EvalBatch.size
             )
@@ -286,6 +509,9 @@ class LevanterHarnessLM(LM):
             padding_count, batch_tokens = _get_padding_count(batch, self.tokenizer.pad_token_id)
 
             out_ids, out_lls, out_correct = self.leader.dispatch_loglikelihood(batch)
+
+            # Increment step after processing batch
+            self._current_step += 1
 
             out_ids = np.array(out_ids.array)
             out_lls = np.array(out_lls.array)
@@ -313,6 +539,9 @@ class LevanterHarnessLM(LM):
             )
             pbar.update(batch_tokens)
 
+        # Ensure profiler is stopped if it was started
+        self._stop_profiler_if_needed()
+
         missing_points = np.where(~covered_points)[0]
         assert len(missing_points) == 0, f"Missing points: {missing_points}"
 
@@ -321,11 +550,301 @@ class LevanterHarnessLM(LM):
 
         return result
 
+    def tok_encode(
+        self,
+        string: Union[str, List[str]],
+        left_truncate_len: int | None = None,
+        add_special_tokens: bool = False,
+        truncation: bool = False,
+    ) -> Union[List[int], List[List[int]]]:
+        """
+        Tokenize a string or list of strings.
+
+        Args:
+            string: The string(s) to tokenize.
+            left_truncate_len: If provided, left-truncate the encoded tokens to this length.
+            add_special_tokens: Whether to add special tokens during tokenization.
+            truncation: Whether to enable tokenizer truncation.
+
+        Returns:
+            Token IDs as a list (for single string) or list of lists (for multiple strings).
+        """
+        if not add_special_tokens:
+            add_special_tokens = False
+        encoding: Union[List[List[int]], List[int]] = self.tokenizer(
+            string,
+            add_special_tokens=add_special_tokens,
+            truncation=truncation,
+            return_attention_mask=False,
+        ).input_ids
+
+        # left-truncate the encoded context to be at most `left_truncate_len` tokens long
+        if left_truncate_len:
+            if not isinstance(string, str):
+                encoding = [enc[-left_truncate_len:] for enc in encoding]  # type: ignore
+            else:
+                encoding = encoding[-left_truncate_len:]
+
+        return encoding
+
+    def _process_and_tokenize_stop_sequences(self, until: Optional[List[str]], eos: str) -> Optional[NamedArray]:
+        """
+        Process and tokenize stop sequences for generation.
+
+        Args:
+            until: List of stop sequences, if any
+            eos: End-of-sequence token string
+
+        Returns:
+            NamedArray with tokenized stop sequences, or None if no valid stop sequences
+        """
+        if not until:
+            return None
+
+        # Process stop sequences to ensure EOS is included
+        processed_until = handle_stop_sequences(until, eos=eos)
+
+        if not processed_until:
+            return None
+
+        # Tokenize all stop sequences
+        all_stop_tokens = []
+        for stop_seq in processed_until:
+            stop_ids_list = self.tok_encode(stop_seq, add_special_tokens=False)
+            if len(stop_ids_list) > 0:
+                all_stop_tokens.append(stop_ids_list)
+
+        if not all_stop_tokens:
+            return None
+
+        # Find the maximum length for padding
+        max_len = max(len(tokens) for tokens in all_stop_tokens)
+
+        # Left pad all sequences to the same length
+        padded_tokens = []
+        for tokens in all_stop_tokens:
+            padding_needed = max_len - len(tokens)
+            padded = [INVALID] * padding_needed + tokens
+            padded_tokens.append(padded)
+
+        # Convert to named array with proper dimensions
+        stop_tokens_array = jnp.asarray(padded_tokens, dtype=jnp.int32)
+        return haliax.named(stop_tokens_array, ("stop_seq", "position"))
+
     def loglikelihood_rolling(self, requests) -> List[Tuple[float]]:
         raise NotImplementedError()
 
     def generate_until(self, requests) -> List[str]:
-        raise NotImplementedError()
+        # Error out on multihost JAX - Engine doesn't support it yet
+        if jax.process_count() > 1:
+            raise NotImplementedError(
+                "InferenceEngine does not yet support multihost JAX. " "Please use a single host for generation tasks."
+            )
+
+        # print(f'len(requests)={len(requests)}')
+
+        # Implement simple generation using InferenceEngine.
+        # requests: list[Instance] where args[0] = prompt, args[1] may be stop strings (list[str])
+        # kwargs may include max_gen_toks, temperature, n (n_generations), seed
+        if self.tokenizer.pad_token_id is None:
+            logger.warning("No pad token set. Setting to eos token.")
+            self.tokenizer.pad_token_id = self.tokenizer.eos_token_id
+
+        # Require a model with paged decode support
+        if not hasattr(self.leader.model, "initial_cache") or not hasattr(self.leader.model, "decode"):
+            raise NotImplementedError(
+                "generate_until requires a model with paged decode support (initial_cache/decode)."
+            )
+
+        # Extract prompts and per-request params following vLLM pattern
+        # batch tokenize contexts
+        context, all_gen_kwargs = zip(*(req.args for req in requests))
+        processed_kwargs_list: list[dict] = []
+
+        for i, gen_kwargs in enumerate(all_gen_kwargs):
+            # print(f'{gen_kwargs=}')
+
+            # Copy and process generation kwargs
+            processed_gen_kwargs = gen_kwargs.copy()
+
+            # Apply defaults from generation_kwargs (user config) first
+            for key, value in self.generation_kwargs.items():
+                processed_gen_kwargs.setdefault(key, value)
+
+            # Standardize kwargs using our _modify_gen_kwargs method
+            processed_gen_kwargs = self._modify_gen_kwargs(processed_gen_kwargs)
+            processed_kwargs_list.append(processed_gen_kwargs)
+
+        # Tokenize prompts and compute capacity needs
+        prompt_token_lists: list[list[int]] = self.tok_encode(context, add_special_tokens=False)  # type: ignore
+
+        # Truncate from left if needed to fit model max length, accounting for generation tokens
+        max_length = self.EvalPos.size
+        for i, (toks, gen_kwargs) in enumerate(zip(prompt_token_lists, processed_kwargs_list)):
+            # Reserve space for generation tokens
+            max_gen_toks = gen_kwargs["max_gen_toks"]
+            max_ctx_len = max_length - max_gen_toks
+
+            if len(toks) > max_ctx_len:
+                overflow = len(toks) - max_ctx_len
+                logger.warning(f"Prompt {i} too long ({len(toks)}). Truncating left by {overflow}.")
+                prompt_token_lists[i] = toks[-max_ctx_len:]
+
+        # Process stop sequences for each request individually
+        # Get EOS token for stop sequence handling
+        eos = self.tokenizer.decode(self.eot_token_id)
+
+        # Process stop sequences and tokenize them for each request
+        for gen_kwargs in processed_kwargs_list:
+            gen_kwargs["stop_tokens"] = self._process_and_tokenize_stop_sequences(gen_kwargs.get("until"), eos)
+
+        # Calculate max stop sequences and tokens from all requests
+        max_stop_seqs = 4
+        max_stop_tokens = 16
+
+        for gen_kwargs in processed_kwargs_list:
+            stop_tokens = gen_kwargs.get("stop_tokens")
+            if stop_tokens is not None:
+                # stop_tokens is a NamedArray with shape (stop_seq, position)
+                num_stop_seqs = stop_tokens.shape["stop_seq"]
+                num_stop_tokens = stop_tokens.shape["position"]
+                max_stop_seqs = max(max_stop_seqs, num_stop_seqs)
+                max_stop_tokens = max(max_stop_tokens, num_stop_tokens)
+
+        # Taken from: config/sampler/sample_llama8b.yaml
+        engine_cfg = InferenceEngineConfig(
+            max_stop_seqs=max_stop_seqs,
+            max_stop_tokens=max_stop_tokens,
+            max_seq_len=max_length,
+            max_seqs=256,
+            page_size=8,
+            compute_dtype=jnp.bfloat16,
+        )
+
+        engine = InferenceEngine.from_model_with_config(
+            model=self.leader.model, tokenizer=self.tokenizer, config=engine_cfg
+        )
+
+        # Build generation requests
+        base_key = jrandom.PRNGKey(engine_cfg.seed)
+        gen_requests: list[GenRequest] = []
+        for i, (toks, gen_kwargs) in enumerate(zip(prompt_token_lists, processed_kwargs_list)):
+            # Extract parameters from processed kwargs
+            max_gen_toks = gen_kwargs["max_gen_toks"]
+            temperature = gen_kwargs["temperature"]
+            n_generations = gen_kwargs["n"]
+            seed = gen_kwargs.get("seed")
+            stop_tokens = gen_kwargs.get("stop_tokens")
+            # print(f'{temperature=}')
+            # print(f'{stop_tokens=}')
+
+            # Create sequence decoding parameters
+            seq_params = SeqDecodingParams(
+                max_num_tokens=jnp.array(len(toks) + max_gen_toks, dtype=jnp.int32),
+                stop_tokens=stop_tokens,
+                temperature=jnp.array(temperature, dtype=jnp.float32),
+                key=jrandom.fold_in(base_key if seed is None else jrandom.PRNGKey(seed), i),
+            )
+            gen_requests.append(
+                GenRequest(
+                    prompt_tokens=list(map(int, toks)),
+                    request_id=i,
+                    decode_params=seq_params,
+                    n_generations=int(n_generations),
+                    enable_logprobs=False,
+                )
+            )
+
+        # Create step callback for profiling decode iterations
+        def decode_step_callback(iteration: int):
+            """Called at each decode iteration in the engine."""
+            # Use the iteration number as the step for profiling
+            saved_step = self._current_step
+            self._current_step = iteration
+            self._handle_profiler_step()
+            self._current_step = saved_step
+
+        # Pass the callback to the engine if profiling is enabled
+        step_callback = decode_step_callback if self.profiler_config.enabled else None
+        result = engine.generate(gen_requests, step_callback=step_callback)
+
+        # Decode first generation per request (LM Harness expects one string per request)
+        outputs: list[str] = []
+        output_idx = 0
+        for i, (toks, gen_kwargs) in enumerate(zip(prompt_token_lists, processed_kwargs_list)):
+            # Consume one sequence output per request
+            if output_idx < len(result.tokens):
+                full_tokens = result.tokens[output_idx]
+                # Engine tokens are generated tokens only (prompt not included)
+                text = self.tokenizer.decode(full_tokens, skip_special_tokens=True)
+
+                # Post-process the generated text using the imported utility function
+                text = postprocess_generated_text(
+                    text, gen_kwargs.get("until"), None  # think_end_token - could be made configurable if needed
+                )
+                outputs.append(text)
+                output_idx += 1  # consume one generation per request
+            else:
+                text = ""
+                outputs.append(text)
+
+            current_task = getattr(self, "_current_task", "generation_task")
+            bucket = self._prepare_bucket(current_task)
+            if bucket is not None:
+                prompt_text = self.tokenizer.decode(toks, skip_special_tokens=False)
+                bucket.append(
+                    {
+                        "prompt": prompt_text,
+                        "generation": text,
+                    }
+                )
+        # print(f'{outputs=}')
+
+        # Stop profiler if it was started during generation
+        self._stop_profiler_if_needed()
+
+        return outputs
+
+    @staticmethod
+    def _modify_gen_kwargs(kwargs: dict) -> dict:
+        """
+        Modify generation kwargs to standardize parameters, similar to vLLM implementation.
+        """
+        # Handle temperature
+        if "temperature" in kwargs and kwargs["temperature"] is not None:
+            kwargs["temperature"] = max(0.0, float(kwargs["temperature"]))
+        else:
+            kwargs.setdefault("temperature", 0.0)
+
+        # Handle do_sample parameter like vLLM does
+        do_sample = kwargs.pop("do_sample", None)
+        if do_sample is False:
+            if kwargs["temperature"] == 0.0:
+                logger.debug("Got `do_sample=False` with temperature 0.0, ensuring deterministic sampling")
+            elif kwargs["temperature"] > 0.0:
+                raise ValueError(
+                    f"Conflicting parameters: do_sample=False but temperature={kwargs['temperature']} > 0.0. "
+                    f"For deterministic sampling, set temperature=0.0 or remove do_sample=False."
+                )
+
+        # Handle max_gen_toks parameter
+        if "max_gen_toks" in kwargs and kwargs["max_gen_toks"] is not None:
+            kwargs["max_gen_toks"] = int(kwargs["max_gen_toks"])
+        else:
+            kwargs.setdefault("max_gen_toks", 256)
+
+        # Handle n generations parameter
+        if "n" in kwargs and kwargs["n"] is not None:
+            kwargs["n"] = int(kwargs["n"])
+        else:
+            kwargs.setdefault("n", 1)
+
+        # Handle seed parameter
+        if "seed" in kwargs and kwargs["seed"] is not None:
+            kwargs["seed"] = int(kwargs["seed"])
+        # Note: seed can remain None, which is valid
+
+        return kwargs
 
 
 @dataclass(frozen=True)
@@ -379,10 +898,32 @@ class TaskConfig:
 class LmEvalHarnessConfig:
     task_spec: list[TaskConfig | str]
     max_examples: int | None = None
-    max_eval_length: int | None = None
+    max_length: int | None = None
     log_samples: bool = False
     bootstrap_iters: int = 0
     apply_chat_template: bool = False
+    fewshot_as_multiturn: bool = False
+    confirm_run_unsafe_code: bool = True
+    sample_logging: SampleLoggingConfig = dataclasses.field(default_factory=SampleLoggingConfig)
+    generation_kwargs: dict = dataclasses.field(
+        default_factory=lambda: {"max_gen_toks": 256, "temperature": 0.0, "n": 1, "seed": None}
+    )
+    """
+    Default generation parameters for text generation tasks.
+
+    Supported parameters:
+    - max_gen_toks: Maximum number of tokens to generate (default: 256)
+    - temperature: Sampling temperature, 0.0 for deterministic (default: 0.0)
+    - n: Number of completions to generate per prompt (default: 1)
+    - seed: Random seed for generation, None for random (default: None)
+
+    These can be overridden on a per-request basis by the evaluation harness.
+    """
+
+    @property
+    def max_gen_toks(self) -> int:
+        """Backward compatibility property for max_gen_toks."""
+        return self.generation_kwargs.get("max_gen_toks", 256)
 
     def to_task_spec(self) -> list[str | dict]:
         return [task.to_dict() if isinstance(task, TaskConfig) else task for task in self.task_spec]
@@ -515,6 +1056,7 @@ class EvalHarnessMainConfig:
     checkpoint_is_hf: bool = False
     """If True, the checkpoint is a HuggingFace checkpoint. Otherwise, it is a Levanter checkpoint."""
     apply_chat_template: bool = False
+    fewshot_as_multiturn: bool = False
     """
     Whether or not to apply the chat template this model was trained with before running inference
     """
@@ -537,9 +1079,13 @@ def run_lm_eval_harness(
     EvalBatch,
     axis_resources,
     mp: jmp.Policy | None,
+    profiler_config: ProfilerConfig | None = None,
 ) -> dict | None:
     """
     Run the LM Eval Harness on the given model and tasks.
+
+    Args:
+        profiler_config: Optional ProfilerConfig for profiling during evaluation
 
     Returns:
         If running on process 0, returns the outputs of the LM Eval Harness with the following extra keys.
@@ -548,7 +1094,9 @@ def run_lm_eval_harness(
     """
     tasks_to_run = config.to_task_dict()
 
-    outputs = _actually_run_eval_harness(config, model, tasks_to_run, tokenizer, EvalBatch, axis_resources, mp)
+    outputs = _actually_run_eval_harness(
+        config, model, tasks_to_run, tokenizer, EvalBatch, axis_resources, mp, profiler_config
+    )
 
     return outputs
 
@@ -561,6 +1109,7 @@ def _actually_run_eval_harness(
     EvalBatch: haliax.Axis,
     axis_resources: ResourceMapping,
     mp: jmp.Policy | None,
+    profiler_config: ProfilerConfig | None = None,
 ) -> dict | None:
     """
     Actually run the LM Eval Harness on the given model and tasks. This is a separate function so that it can be used
@@ -572,12 +1121,12 @@ def _actually_run_eval_harness(
 
     """
     max_examples = config.max_examples
-    max_eval_length = config.max_eval_length
+    max_length = config.max_length
 
-    EvalPos = model.Pos if max_eval_length is None else model.Pos.resize(max_eval_length)
+    EvalPos = model.Pos if max_length is None else model.Pos.resize(max_length)
     num_parameters = levanter.utils.jax_utils.parameter_count(model)
     logger.info(
-        f"Evaluating with max eval length {EvalPos.size} and batch size {EvalBatch.size}. There are"
+        f"Evaluating with max length {EvalPos.size} and batch size {EvalBatch.size}. There are"
         f" {num_parameters} parameters in the model."
     )
     logger.info("Running eval harness...")
@@ -590,7 +1139,9 @@ def _actually_run_eval_harness(
         tokenizer,
         mp,
         max_packed_segments=64,
-        apply_chat_template=config.apply_chat_template,
+        generation_kwargs=config.generation_kwargs,
+        sample_logging_config=config.sample_logging,
+        profiler_config=profiler_config,
     )
 
     if jax.process_index() == 0:
@@ -599,6 +1150,10 @@ def _actually_run_eval_harness(
 
         # eval_harness only sets seeds in simple_evaluate, which we can't use (I think?)
         tasks_to_run = _adjust_config(tasks_to_run, 0)
+
+        # Clear any previous sample outputs
+        harness.clear_sample_outputs()
+
         with set_global_rng_seeds(0):
             outputs = evaluator.evaluate(
                 harness,
@@ -606,11 +1161,27 @@ def _actually_run_eval_harness(
                 limit=max_examples,
                 log_samples=config.log_samples,
                 bootstrap_iters=config.bootstrap_iters,
+                apply_chat_template=config.apply_chat_template,
+                fewshot_as_multiturn=config.fewshot_as_multiturn,
+                confirm_run_unsafe_code=config.confirm_run_unsafe_code,
             )
+
         worker.stop()
 
         averages = _compute_averages(outputs)
         outputs["averages"] = averages
+
+        # Get the collected sample outputs and add them to the results
+        sample_outputs = harness.get_sample_outputs()
+        if config.sample_logging.should_log() and sample_outputs:
+            # Add outputs to each benchmark in results
+            for task_name in outputs.get("results", {}):
+                # Get all sample outputs for this task (since we don't track individual tasks yet)
+                all_samples = []
+                for samples in sample_outputs.values():
+                    all_samples.extend(samples)
+                if all_samples:
+                    outputs["results"][task_name]["outputs"] = all_samples
 
         return outputs
     else:
@@ -669,12 +1240,14 @@ def _compute_averages(outputs):
 
 def run_eval_harness_main(config: EvalHarnessMainConfig):
     config.trainer.initialize()
+    # Ensure __main__ logger is at INFO level for profiler messages
+    logger.setLevel(logging.INFO)
     tokenizer = config.the_tokenizer
 
     compute_axis_mapping = config.trainer.compute_axis_mapping
     parameter_axis_mapping = config.trainer.parameter_axis_mapping
 
-    with config.trainer.device_mesh, hax.axis_mapping(parameter_axis_mapping):
+    with config.trainer.use_device_mesh(), hax.axis_mapping(parameter_axis_mapping):
         key = jax.random.PRNGKey(0)
 
         vocab_size = len(tokenizer)
@@ -700,6 +1273,20 @@ def run_eval_harness_main(config: EvalHarnessMainConfig):
 
         model = typing.cast(LmHeadModel, inference_mode(model, True))
 
+        # Set up profiler configuration if enabled
+        profiler_config = None
+        if config.trainer.profiler:
+            # Get the run_id that was set during initialize()
+            run_id = config.trainer._maybe_set_id()
+            profile_path = config.trainer.log_dir / run_id / "profiler"
+            profiler_config = ProfilerConfig(
+                enabled=True,
+                start_step=config.trainer.profiler_start_step,
+                num_steps=config.trainer.profiler_num_steps,
+                profile_path=str(profile_path),
+                perfetto_link=config.trainer.profiler_perfetto_link,
+            )
+
         logger.info("Running LM eval harness....")
         outputs = run_lm_eval_harness(
             config.eval_harness,
@@ -708,6 +1295,7 @@ def run_eval_harness_main(config: EvalHarnessMainConfig):
             config.EvalBatch,
             axis_resources=compute_axis_mapping,
             mp=config.trainer.mp,
+            profiler_config=profiler_config,
         )
 
         logger.info("Finished running LM eval harness")
@@ -722,12 +1310,12 @@ def run_eval_harness_main(config: EvalHarnessMainConfig):
             # log the results as json
             logger.info("uploading artifacts...")
             with open("lm_eval_harness_results.json", "w") as f:
-                json.dump(outputs, f, indent=2)
+                json.dump(outputs, f, indent=2, default=_json_default)
                 f.flush()
                 f_path = f.name
                 levanter.tracker.current_tracker().log_artifact(f_path, name="lm_eval_harness_results")
 
-            print(json.dumps(outputs, indent=2), flush=True)
+            print(json.dumps(outputs, indent=2, default=_json_default), flush=True)
 
         return outputs
 
@@ -754,7 +1342,30 @@ def log_report_to_tracker(prefix: str, report: dict, tracker: Optional[levanter.
 
                 to_log[f"{prefix}/averages/{metric_name}"] = metric_value
 
-    tracker.log(to_log, step=None)
+    if to_log:
+        tracker.log(to_log, step=None)
+
+
+def _json_default(value):
+    """
+    Provide a best-effort JSON serialization for objects returned by the eval harness.
+    """
+    if dataclasses.is_dataclass(value):
+        return dataclasses.asdict(value)
+
+    if isinstance(value, set):
+        return list(value)
+
+    if hasattr(value, "to_dict") and callable(value.to_dict):
+        try:
+            return value.to_dict()
+        except Exception:
+            pass
+
+    if hasattr(value, "__dict__"):
+        return vars(value)
+
+    return repr(value)
 
 
 def lm_eval_harness(config: LmEvalHarnessConfig, tokenizer, EvalBatch, axis_resources, mp: jmp.Policy | None):
@@ -766,6 +1377,8 @@ def lm_eval_harness(config: LmEvalHarnessConfig, tokenizer, EvalBatch, axis_reso
 
         model = step.eval_model
         logger.info("Running eval harness...")
+        # Note: profiler_config is None here since this is used as a callback during training
+        # and the trainer handles profiling separately
         outputs = _actually_run_eval_harness(
             config,
             model,
@@ -774,6 +1387,7 @@ def lm_eval_harness(config: LmEvalHarnessConfig, tokenizer, EvalBatch, axis_reso
             EvalBatch,
             axis_resources,
             mp,
+            profiler_config=None,
         )
         logger.info("Finished running eval harness.")
 
@@ -786,7 +1400,7 @@ def lm_eval_harness(config: LmEvalHarnessConfig, tokenizer, EvalBatch, axis_reso
             with tempfile.NamedTemporaryFile("w", delete=False, suffix=".json") as f:
                 import json
 
-                json.dump(outputs, f)
+                json.dump(outputs, f, default=_json_default)
                 f.flush()
                 levanter.tracker.current_tracker().log_artifact(
                     f.name, name=f"lm_eval_harness_results.{step.step}.json", type="lm_eval_output"
@@ -816,20 +1430,12 @@ def _adjust_config(task_dict, fewshot_random_seed=0):
 
 
 def _iterate_tokenized_requests(
-    requests: list[Instance], tokenizer: HfTokenizer, max_len: int, batch_size: int, apply_chat_template: bool = False
+    requests: list[Instance], tokenizer: HfTokenizer, max_length: int, batch_size: int
 ) -> Iterator[PromptCompletion]:
     """
     Tokenize the requests and yield them as PromptCompletions, for packing into LmExamples.
     """
-    if apply_chat_template:
-        contexts = [
-            tokenizer.apply_chat_template(
-                [{"role": "user", "content": request.args[0]}], tokenize=False, add_generation_prompt=True
-            )
-            for request in requests
-        ]
-    else:
-        contexts = [request.args[0] for request in requests]
+    contexts = [request.args[0] for request in requests]
 
     completions = [request.args[1] for request in requests]
 
@@ -852,11 +1458,11 @@ def _iterate_tokenized_requests(
 
             context_enc_len = len(context_enc)
 
-            if len(all_enc) > max_len:
+            if len(all_enc) > max_length:
                 logger.warning(f"Request {i} is too long. Truncating.")
                 # Truncate from the left
-                context_enc_len = len(context_enc) - (len(all_enc) - max_len)
-                all_enc = all_enc[-max_len:]
+                context_enc_len = len(context_enc) - (len(all_enc) - max_length)
+                all_enc = all_enc[-max_length:]
                 if context_enc_len < 0:
                     context_enc_len = 0
                     logger.warning("Prompt length is negative after truncation. Setting to 0.")
@@ -868,11 +1474,8 @@ def _pack_requests(
     tokenizer: HfTokenizer,
     Pos: hax.Axis,
     max_pack_size: int,
-    apply_chat_template: bool = False,
 ) -> list[LmExample]:
-    packed_iterator = _iterate_tokenized_requests(
-        requests, tokenizer, Pos.size, batch_size=128, apply_chat_template=apply_chat_template
-    )
+    packed_iterator = _iterate_tokenized_requests(requests, tokenizer, Pos.size, batch_size=128)
     # TODO: use a better packing algorithm?
     return greedy_pack_prompt_completions(
         Pos,
