@@ -1,3 +1,6 @@
+# Copyright 2025 The Levanter Authors
+# SPDX-License-Identifier: Apache-2.0
+
 __all__ = [
     # attention
     "Attention",
@@ -6,7 +9,6 @@ __all__ = [
     "AttentionConfig",
     "AttentionMask",
     "dot_product_attention",
-    "dot_product_attention_with_sink",
     # normalization
     "LayerNormConfig",
     "LayerNormConfigBase",
@@ -20,6 +22,5 @@ from .attention import (
     AttentionMask,
     AttentionWithSink,
     dot_product_attention,
-    dot_product_attention_with_sink,
 )
 from .normalization import LayerNormConfig, LayerNormConfigBase, RmsNormConfig
