@@ -1,3 +1,6 @@
+# Copyright 2025 The Levanter Authors
+# SPDX-License-Identifier: Apache-2.0
+
 # cf https://github.com/lucidrains/flash-attention-jax
 # cf https://tridao.me/publications/flash2/flash2.pdf
 # cf https://arxiv.org/pdf/2205.14135.pdf
@@ -31,6 +34,7 @@ def flash_attention(
     v: hax.NamedArray,
     mask: Optional[AttentionMask | hax.NamedArray] = None,
     bias: Optional[hax.NamedArray] = None,
+    attn_sink: Optional[hax.NamedArray] = None,
     *,
     dropout: float = 0.0,
     inference: bool,
@@ -44,12 +48,11 @@ def flash_attention(
     """
     Crappy pure-jax Flash Attention impl, vaguely following the v2 paper.
 
-    Args:
-        Key: axis of key dim.
+    - If `block_size` doesn't divide Q/K, this falls back to the VANILLA path.
+      When `attn_sink` is provided, that fallback is forced to VANILLA to avoid recursion.
     """
     if not inference and dropout > 0 and key is None:
         raise ValueError("key must be provided for training")
-
     if dropout < 0 or dropout > 1:
         raise ValueError(f"invalid dropout {dropout}")
 
@@ -65,9 +68,9 @@ def flash_attention(
     KPos = k.resolve_axis(KPos)
 
     if QPos.size < block_size or KPos.size < block_size:
-        from levanter.layers.attention import simple_attention_with_dropout
+        from levanter.layers.attention import dot_product_attention
 
-        return simple_attention_with_dropout(
+        return dot_product_attention(
             QPos,
             KPos,
             Key,
@@ -76,10 +79,17 @@ def flash_attention(
             v,
             mask=mask,
             bias=bias,
+            attention_dtype=dtype,
+            precision=precision,
+            use_flash=False,  # force VANILLA
+            attn_backend=None,
+            flash_block_size=block_size,
             dropout=dropout,
+            logits_soft_cap=logits_soft_cap,
+            scaling_factor=scaling_factor,
             inference=inference,
             prng=key,
-            logits_soft_cap=logits_soft_cap,
+            attn_sink=attn_sink,
         )
 
     if scaling_factor is None:
@@ -89,7 +99,7 @@ def flash_attention(
     q = q * scaling_factor
 
     return _flash_attention(
-        (q, k, v),
+        (q, k, v, attn_sink),
         QPos,
         KPos,
         Key,
@@ -106,7 +116,7 @@ def flash_attention(
 
 @equinox.filter_custom_vjp
 def _flash_attention(
-    qkv: Tuple[hax.NamedArray, hax.NamedArray, hax.NamedArray],
+    qkv_sink: Tuple[hax.NamedArray, hax.NamedArray, hax.NamedArray, Optional[hax.NamedArray]],
     QPos: hax.Axis,
     KPos: hax.Axis,
     Key: hax.Axis,
@@ -122,7 +132,7 @@ def _flash_attention(
 ) -> hax.NamedArray:
     return _flash_attention_forward(
         None,
-        qkv,
+        qkv_sink,
         QPos,
         KPos,
         Key,
@@ -140,7 +150,7 @@ def _flash_attention(
 @named_call
 def _flash_attention_forward(
     ignore,
-    qkv,
+    qkv_sink,
     QPos: hax.Axis,
     KPos: hax.Axis,
     Key: hax.AxisSelector,
@@ -155,7 +165,7 @@ def _flash_attention_forward(
     logits_soft_cap: Optional[float],
 ):
     del ignore
-    q, k, v = qkv
+    q, k, v, attn_sink = qkv_sink
     if QPos.size % block_size != 0:
         raise ValueError(f"q axis size {q.axis_size(QPos)} is not a multiple of {block_size}")
     if KPos.size % block_size != 0:
@@ -165,6 +175,7 @@ def _flash_attention_forward(
     Tr = QPos.size // block_size
     Tc = KPos.size // block_size
 
+    # Row axes: everything in q except (QPos, Key)
     q_batch_axes: Tuple[hax.Axis, ...] = hax.eliminate_axes(q.axes, (QPos, Key))
 
     # output variables: O is the attention output, ell is the per-position log normalizer
@@ -175,6 +186,11 @@ def _flash_attention_forward(
     ell = hax.auto_sharded(ell)
 
     is_causal = isinstance(mask, AttentionMask) and mask.is_causal
+    # Non-zero offsets (dynamic or otherwise) are not supported in this implementation
+    if isinstance(mask, AttentionMask) and mask.is_causal and mask.causal_offset is not None:
+        raise NotImplementedError("Non-zero causal offsets are not implemented for flash attention")
+
+    # i refers to the block index in Q, j refers to the block index in K.
 
     @named_call
     def do_o_block(state):
@@ -185,9 +201,22 @@ def _flash_attention_forward(
 
         # Step 2: init O_i = 0, sumexp_i = 0, max_i = -inf
         o_i = o[QPos, ds.block(i, block_size)]
+
         QPosBlock = QPos.resize(block_size)
-        sumexp_i = hax.zeros(q_batch_axes + (QPosBlock,), q.dtype)
-        max_i = hax.full(q_batch_axes + (QPosBlock,), -jnp.inf, q.dtype)
+        row_axes_block: Tuple[hax.Axis, ...] = q_batch_axes + (QPosBlock,)
+
+        sumexp_i = hax.zeros(row_axes_block, q.dtype)
+        max_i = hax.full(row_axes_block, -jnp.inf, q.dtype)
+
+        if attn_sink is not None:
+            sink_prefix = attn_sink
+            for ax in q_batch_axes:
+                sink_prefix = sink_prefix.broadcast_axis(ax) if ax not in sink_prefix.axes else sink_prefix
+            sink_block = sink_prefix.broadcast_axis(QPosBlock).astype(q.dtype)
+            # Ensure axis order matches row_axes_block
+            sink_block = sink_block.rearrange(row_axes_block)
+            max_i = sink_block
+            sumexp_i = hax.ones(row_axes_block, q.dtype)
 
         @named_call
         def do_qk_block(state):
@@ -222,25 +251,35 @@ def _flash_attention_forward(
                 attn_ij = hax.where(mask_ij, attn_ij, -1e10)
 
             if dropout > 0 and not inference:
-                attn_ij = hax.nn.dropout(attn_ij, dropout, inference=False, key=jax.random.fold_in(key, i * Tc + j))
+                attn_ij = hax.nn.dropout(
+                    attn_ij,
+                    dropout,
+                    inference=False,
+                    key=jax.random.fold_in(key, i * Tc + j),
+                )
 
             # Step 9: Compute m_i^j = max(m_i^{j-1}, rowmax(S_i^j)), P_i^j = exp(S_i^j - m_i^j),
             # ...    l_i^j = exp(m_i^{j-1} - m_i^j) + rowsum(P_i^j)
             max_i = hax.maximum(old_max_i, hax.max(attn_ij, axis=KPos.name))
             P_ij = hax.exp(attn_ij - max_i)
+            exp_diff = hax.exp(old_max_i - max_i).rearrange(row_axes_block)
 
-            exp_diff = hax.exp(old_max_i - max_i)
-            sumexp_i = exp_diff * sumexp_i + hax.sum(P_ij, axis=KPos.name)
+            rowsum = hax.sum(P_ij, axis=KPos.name).rearrange(row_axes_block)
+            sumexp_i = exp_diff * sumexp_i + rowsum
 
             # Step 10: Compute O_i = diag(exp(m_i^{j-1} - m_i^j) O_i + P_i^j V_j
-            o_i = exp_diff * o_i + hax.dot(P_ij, v_j, axis=KPos.name)
+            o_term = hax.dot(P_ij, v_j, axis=KPos.name).rearrange(o_i.axes)
+            o_i = exp_diff * o_i + o_term
 
             return (i, j + 1, o_i, q_i, sumexp_i, max_i)
 
+        # For causal (offset==0), limit key blocks up to the diagonal
         j_end = jnp.minimum(i + 1, Tc) if is_causal else Tc
 
         _, _, o_i, _, sumexp_i, max_i = jax.lax.while_loop(
-            lambda state: state[1] < j_end, do_qk_block, (i, 0, o_i, q_i, sumexp_i, max_i)
+            lambda state: state[1] < j_end,
+            do_qk_block,
+            (i, 0, o_i, q_i, sumexp_i, max_i),
         )
 
         # Step 12: compute O_i = diag(\ell_i^{Tc})^{-1} O_i^{Tc}
@@ -263,7 +302,7 @@ def _flash_attention_backward(
     residuals,
     grad_in: hax.NamedArray,
     ignore,
-    qkv,
+    qkv_sink,
     QPos: hax.Axis,
     KPos: hax.Axis,
     Key: hax.AxisSelector,
@@ -279,7 +318,7 @@ def _flash_attention_backward(
 ):
     del ignore
     O, L = residuals
-    q, k, v = qkv
+    q, k, v, attn_sink = qkv_sink
     dO = grad_in
 
     Tr = QPos.size // block_size
@@ -321,7 +360,12 @@ def _flash_attention_backward(
             attn_ij = hax.dot(q_i, k_j, precision=precision, axis=Key)
 
             if dropout > 0 and not inference:
-                attn_ij = hax.nn.dropout(attn_ij, dropout, inference=False, key=jax.random.fold_in(key, i * Tc + j))
+                attn_ij = hax.nn.dropout(
+                    attn_ij,
+                    dropout,
+                    inference=False,
+                    key=jax.random.fold_in(key, i * Tc + j),
+                )
 
             if bias is not None:
                 bias_ij = bias[QPos, ds.block(i, block_size), KPos, ds.block(j, block_size)]
@@ -337,7 +381,12 @@ def _flash_attention_backward(
             p_ij = hax.exp(attn_ij - L_i)
 
             if dropout > 0 and not inference:
-                p_ij = hax.nn.dropout(p_ij, dropout, inference=False, key=jax.random.fold_in(key, i * Tc + j))
+                p_ij = hax.nn.dropout(
+                    p_ij,
+                    dropout,
+                    inference=False,
+                    key=jax.random.fold_in(key, i * Tc + j),
+                )
 
             dP_ij = hax.dot(dO_i, v_j, axis=Key)
             dAttn_ij = p_ij * (dP_ij - D_i)
@@ -346,7 +395,6 @@ def _flash_attention_backward(
             dV_ji = hax.dot(p_ij, dO_i, axis=QPos.name).astype(dV_j.dtype)
             dK_ji = hax.dot(dAttn_ij, q_i, axis=QPos.name).astype(dK_j.dtype)
 
-            # GQA-specific: eliminate unnecessary axes (e.g. 'q_heads_per_group')
             unnecessary_axes = hax.eliminate_axes(dV_ji.axes, _strip_sizes(v.axes))
             dV_ji = hax.sum(dV_ji, unnecessary_axes)
             dK_ji = hax.sum(dK_ji, unnecessary_axes)
@@ -373,7 +421,23 @@ def _flash_attention_backward(
 
     # dQ, (dK, dV) = hax.scan(do_kv_block, Tc)(dQ, jnp.arange(Tc.size))
     j, dQ, dK, dV = jax.lax.while_loop(lambda state: state[0] < Tc, do_kv_block, (0, dQ, dK, dV))
-    return dQ.rearrange(q.axes), dK.rearrange(k.axes), dV.rearrange(v.axes)
+
+    if attn_sink is not None:
+        q_batch_axes: Tuple[hax.Axis, ...] = hax.eliminate_axes(q.axes, (QPos, Key))
+        sink_prefix = attn_sink
+        for ax in q_batch_axes:
+            if ax not in sink_prefix.axes:
+                sink_prefix = sink_prefix.broadcast_axis(ax)
+        sink_rows = sink_prefix.broadcast_axis(QPos).astype(L.dtype)
+
+        p_sink = hax.exp(sink_rows - L)
+        dsink_rows = -p_sink * D
+        reduce_axes = tuple(ax for ax in dsink_rows.axes if ax not in attn_sink.axes)
+        dsink = hax.sum(dsink_rows, reduce_axes).astype(attn_sink.dtype)
+    else:
+        dsink = None
+
+    return (dQ.rearrange(q.axes), dK.rearrange(k.axes), dV.rearrange(v.axes), dsink)
 
 
 _flash_attention.def_fwd(_flash_attention_forward)
@@ -386,7 +450,13 @@ def _infer_attention_output_block_shape(QPos, KPos, Key, q_i, k, v):
 
 
 def _materialize_mask_slice(mask, i, j, QPos, KPos, block_size):
-    return materialize_mask(mask, QPos, KPos, q_slice=hax.ds.block(i, block_size), k_slice=hax.ds.block(j, block_size))
+    return materialize_mask(
+        mask,
+        QPos,
+        KPos,
+        q_slice=hax.ds.block(i, block_size),
+        k_slice=hax.ds.block(j, block_size),
+    )
 
 
 def _strip_sizes(axes: AxisSpec) -> AxisSelection:
