@@ -1,8 +1,13 @@
+# Copyright 2025 The Levanter Authors
+# SPDX-License-Identifier: Apache-2.0
+
 import dataclasses
 from dataclasses import dataclass
 from typing import Callable, Dict, Optional, Type, Union
 
 import equinox as eqx
+import haliax.debug
+import jax
 import jax.random as jrandom
 from jaxtyping import PRNGKeyArray
 
@@ -14,8 +19,10 @@ from haliax.nn.scan import ScanCheckpointPolicy, Stacked
 from haliax.state_dict import ModuleWithStateDictSerialization
 
 from levanter.compat.hf_checkpoints import HFCheckpointConverter, HFCompatConfig
+from levanter.inference.page_table import PageBatchInfo, PageTable
 from levanter.layers import LayerNormConfigBase, RmsNormConfig
 from levanter.layers.attention import Attention, AttentionBackend, AttentionConfig, AttentionMask
+from levanter.layers.kv_cache import KvPageCache, ListCache
 from levanter.layers.rotary import DefaultRotaryEmbeddingsConfig, RotaryEmbeddingsConfig
 from levanter.models.lm_model import LmConfig, LmHeadModel
 from levanter.utils.activation import ActivationFunctionEnum
@@ -80,7 +87,7 @@ class LlamaConfig(HFCompatConfig):
     Pos = property(lambda self: Axis(name="position", size=self.seq_len))
     KeyPos = property(lambda self: self.Pos.alias("key_position"))
     Embed = property(lambda self: Axis(name="embed", size=self.hidden_dim))
-    Layers = property(lambda self: Axis(name="layers", size=self.num_layers))
+    Layers = property(lambda self: Axis(name="layer", size=self.num_layers))
     Mlp = property(lambda self: Axis(name="mlp", size=self.intermediate_dim))
 
     def __post_init__(self):
@@ -88,9 +95,7 @@ class LlamaConfig(HFCompatConfig):
             self.num_heads % self.num_kv_heads == 0
         ), f"num_heads={self.num_heads} not divisible by num_kv_heads={self.num_kv_heads}."
 
-    def hf_checkpoint_converter(
-        self, ref_checkpoint: Optional[str] = None
-    ) -> HFCheckpointConverter["LlamaConfig"]:  # type: ignore
+    def hf_checkpoint_converter(self, ref_checkpoint: Optional[str] = None) -> HFCheckpointConverter["LlamaConfig"]:  # type: ignore
         return HFCheckpointConverter(
             self.__class__,
             reference_checkpoint=self.reference_checkpoint if ref_checkpoint is None else ref_checkpoint,
@@ -197,7 +202,7 @@ class LlamaConfig(HFCompatConfig):
     def total_trainable_params(self, vocab_size):
         token_embedding = vocab_size * self.hidden_dim
 
-        head_size = self.hidden_dim // self.num_heads
+        head_size = self.actual_head_size
         q_proj = self.hidden_dim * head_size * self.num_heads
         kv_proj = 2 * self.hidden_dim * head_size * self.num_kv_heads
         o_proj = head_size * self.num_heads * self.hidden_dim
@@ -213,7 +218,8 @@ class LlamaConfig(HFCompatConfig):
         if self.input_embedding_norm:
             transformer += self.hidden_dim
 
-        return transformer + token_embedding * 2  # plus embedding and lm head
+        lm_head = 0 if self.tie_word_embeddings else token_embedding
+        return transformer + token_embedding + lm_head
 
     def attention_config(self) -> AttentionConfig:
         """Convert this LlamaConfig to an AttentionConfig for use with Attention."""
@@ -228,6 +234,13 @@ class LlamaConfig(HFCompatConfig):
             flash_attention_block_size=self.flash_attention_block_size,
             rope=self.rope,
         )
+
+    @property
+    def actual_head_size(self):
+        """Returns the actual head size based on the head_dim or calculated from hidden_dim and num_heads."""
+        if self.head_dim is not None:
+            return self.head_dim
+        return self.hidden_dim // self.num_heads
 
 
 class LlamaMlp(eqx.Module):
@@ -323,6 +336,42 @@ class LlamaDecoderLayer(eqx.Module):
         output = residual + mlp_output
         return output
 
+    @named_call
+    def decode(
+        self,
+        x: NamedArray,
+        kv_cache: KvPageCache,
+        batch_info: PageBatchInfo,
+        pos_ids: NamedArray,
+        *,
+        key=None,
+    ) -> tuple[NamedArray, KvPageCache]:
+        k_attn, k_mlp = maybe_rng_split(key, 2)
+        # self attention and skip connection
+        residual = x
+        x = self.input_layernorm(x)
+        attn_output, kv_cache = self.self_attn.paged_decode(x, kv_cache, batch_info, pos_ids=pos_ids, key=k_attn)
+
+        if self.post_attn_layernorm is not None:
+            attn_output = self.post_attn_layernorm(attn_output)
+        x = residual + attn_output
+
+        # MLP and skip connection
+        residual = x
+        x = self.post_attention_layernorm(x)
+        mlp_output = self.mlp(x, key=k_mlp)
+        if self.post_mlp_layernorm is not None:
+            mlp_output = self.post_mlp_layernorm(mlp_output)
+        output = residual + mlp_output
+        return output, kv_cache
+
+    def initial_cache(self, page_table: PageTable, *, dtype) -> KvPageCache:
+        """
+        Creates an empty page cache for this layer. Note that in order to create a decoder state, you
+        need to couple the KvPageCache to the PageTable's state with a BatchInfo object.
+        """
+        return self.self_attn.empty_page_cache(page_table, dtype=dtype)
+
 
 class LlamaTransformer(eqx.Module):
     config: LlamaConfig = eqx.field(static=True)
@@ -354,6 +403,60 @@ class LlamaTransformer(eqx.Module):
         x = self.norm(x)
 
         return x
+
+    @named_call
+    def decode(
+        self,
+        kv_cache: ListCache[KvPageCache],
+        x: NamedArray,
+        batch_info: PageBatchInfo,
+        pos_ids: NamedArray,
+        *,
+        key=None,
+    ) -> tuple[NamedArray, ListCache[KvPageCache]]:
+        keys = maybe_rng_split(key, self.config.num_layers) if key is not None else None
+
+        # Unfortunately, JAX does not seem to want to intelligently reuse memory here, so we manually unroll the loop
+        # x, kv_cache = self.layers.scan_via(LlamaDecoderLayer.decode)(
+        #     x,
+        #     kv_cache,
+        #     batch_info,
+        #     pos_ids=pos_ids,
+        #     key=keys,
+        # )
+        # x = self.norm(x)
+        # return x, kv_cache
+        caches = list(kv_cache)
+        updated_caches: list[KvPageCache] = []
+
+        for i in range(self.config.num_layers):
+            with jax.named_scope("slice layer"):
+                layer = hax.tree_util.tree_map(lambda l: l["layer", i], self.layers.stacked)  # type: ignore
+            with jax.named_scope("slice cache"):
+                this_cache = caches[i]
+            x, this_cache = layer.decode(
+                x,
+                this_cache,
+                batch_info,
+                pos_ids=pos_ids,
+                key=keys[i] if keys is not None else None,
+            )
+            with jax.named_scope("update cache"):
+                updated_caches.append(this_cache)
+
+        x = self.norm(x)
+
+        return x, ListCache(updated_caches)
+
+    def initial_cache(self, page_table: PageTable, *, dtype) -> ListCache[KvPageCache]:
+        """
+        Creates an empty page cache for this transformer. Note that in order to create a decoder state, you
+        need to couple the KvPageCache to the PageTable's state with a BatchInfo object.
+        """
+        # sadly this is too cute/smart for XLA to handle aliasing correctly
+        # return self.layers.vmap_via(LlamaDecoderLayer.initial_cache)(page_table, dtype=dtype)
+        caches = [layer.initial_cache(page_table, dtype=dtype) for layer in self.layers.unstacked()]
+        return ListCache(caches)
 
 
 class LlamaEmbedding(ModuleWithStateDictSerialization, eqx.Module):
@@ -501,3 +604,58 @@ class LlamaLMHeadModel(ModuleWithStateDictSerialization, LmHeadModel[LlamaConfig
 
     def _state_dict_key_map(self) -> Dict[str, Optional[str]]:
         return {"transformer": "model", "embeddings": None}
+
+    def initial_cache(self, page_table: PageTable, *, dtype) -> ListCache[KvPageCache]:
+        """
+        Creates an initial cache for this model. Note that in order to create a decoder state, you
+        need to couple the KvPageCache to the PageTable's state with a BatchInfo object.
+        """
+        return hax.auto_sharded(self.transformer.initial_cache(page_table, dtype=dtype))
+
+    @named_call
+    def decode(
+        self,
+        input_ids: NamedArray,  # token IDs for *this* step (shape {Pos} or {Batch, Pos})
+        kv_cache: ListCache[KvPageCache],
+        batch_info: PageBatchInfo,
+        pos_ids: NamedArray,
+        *,
+        key=None,
+    ) -> tuple[NamedArray, ListCache[KvPageCache]]:
+        """Run one decode / pre-fill step with an existing paged-KV *state*.
+
+        Parameters
+        ----------
+        input_ids : NamedArray
+            Token IDs for the positions being decoded **this call**.
+        kv_cache : ListCache[KvPageCache]
+            Current paged-KV cache (one per layer). Obtain the initial value via
+            ``self.initial_cache`` and update with the returned *new_state* each step.
+        pos_ids : NamedArray
+            Absolute position IDs matching *input_ids* (negative IDs can mark padding as
+            in the lower-level API).
+        key : jax.random.PRNGKey | None
+            RNG key for dropout etc.  Can be omitted during inference.
+
+        Returns
+        -------
+        logits : NamedArray
+            Logits for the provided tokens (axes match *input_ids* + ``Vocab``).
+        new_state : ListCache[KvPageCache]
+            Updated cache to pass into the next decode call.
+        """
+
+        # Embed the incoming token IDs
+        x = self.embeddings.embed(input_ids)
+
+        # Propagate through the transformer with paged-KV caching
+        k_t = maybe_rng_split(key, 1)[0] if key is not None else None
+        x, new_state = self.transformer.decode(kv_cache, x, batch_info, pos_ids, key=k_t)
+
+        # Project to logits
+        if self.lm_head is not None:
+            logits = self.lm_head(x, key=None)
+        else:
+            logits = self.embeddings.unembed(x)
+
+        return logits, new_state
