@@ -1,6 +1,10 @@
+# Copyright 2025 The Levanter Authors
+# SPDX-License-Identifier: Apache-2.0
+
 import dataclasses
 from dataclasses import dataclass
-from typing import NamedTuple
+from functools import partial
+from typing import NamedTuple, Optional
 
 import chex
 import jax
@@ -32,6 +36,7 @@ class MuonConfig(OptimizerConfig):
     nesterov: bool = True
     backend_steps: int = 5  # Number of steps for Newton-Schulz orthogonalization
     weight_decay: float = 0.0
+    adam_weight_decay: Optional[float] = None
     beta1: float = 0.9
     beta2: float = 0.95
     epsilon: float = 1e-8
@@ -51,7 +56,11 @@ class MuonConfig(OptimizerConfig):
         def optimizer(learning_rate, adam_lr):
             def muon_transform():
                 components = []
-                components.append(scale_with_muon(self.momentum, self.nesterov, self.backend_steps, self.muon_epsilon, self.use_kimi_scaling))
+                components.append(
+                    scale_with_muon(
+                        self.momentum, self.nesterov, self.backend_steps, self.muon_epsilon, self.use_kimi_scaling
+                    )
+                )
                 if self.weight_decay > 0:
                     components.append(optax.add_decayed_weights(self.weight_decay, self.build_weight_decay_mask()))
                 components.append(optax.scale(-learning_rate))
@@ -63,8 +72,9 @@ class MuonConfig(OptimizerConfig):
                 if self.max_grad_norm:
                     components.append(optax.clip_by_global_norm(self.max_grad_norm))
                 components.append(optax.scale_by_adam(self.beta1, self.beta2, self.epsilon))
-                if self.weight_decay > 0:
-                    components.append(optax.add_decayed_weights(self.weight_decay, self.build_weight_decay_mask()))
+                adam_weight_decay = self.adam_weight_decay if self.adam_weight_decay is not None else self.weight_decay
+                if adam_weight_decay > 0:
+                    components.append(optax.add_decayed_weights(adam_weight_decay, self.build_weight_decay_mask()))
                 components.append(optax.scale(-adam_lr))
                 optimizer = optax.chain(*components)
                 return optimizer
@@ -74,11 +84,13 @@ class MuonConfig(OptimizerConfig):
                 "adamw": adamw_transform(),
             }
 
-            return optax.multi_transform(transformations, partial(self.create_mask, use_kimi_scaling = use_kimi_scaling))
+            return optax.multi_transform(
+                transformations, partial(self.create_mask, use_kimi_scaling=self.use_kimi_scaling)
+            )
 
         return optax.inject_hyperparams(optimizer)(learning_rate=learning_rate_schedule, adam_lr=adam_lr_schedule)
 
-    def create_mask(self, params, use_kimi_scaling = True):
+    def create_mask(self, params, use_kimi_scaling=True):
         """
         Creates a mask that labels parameters as 'muon' or 'adamw' based on their
         dimensionality and module path, using AdamW for Embedding and lm_head parameters.
@@ -91,7 +103,9 @@ class MuonConfig(OptimizerConfig):
                 return "adamw"
             elif isinstance(param, Linear):
                 # muon for linear layers
-                assert (param._out_first or use_kimi_scaling) # if we don't use kimi's version of scaling, then we need to assume out_first to ensure we are scaling like Out/In
+                assert (
+                    param._out_first or use_kimi_scaling
+                )  # if we don't use kimi's version of scaling, then we need to assume out_first to ensure we are scaling like Out/In
                 return dataclasses.replace(param, weight="muon", bias="adamw" if param.bias is not None else None)
             else:
                 return "adamw"
@@ -138,7 +152,9 @@ def scale_with_muon(momentum=0.95, nesterov=True, steps=5, muon_eps=1e-8, use_ki
             updated_weight_array = zeropower_via_newtonschulz5(array, steps=steps, eps=muon_eps)
 
             if not use_kimi_scaling:
-                scale = jnp.sqrt(jnp.maximum(1, updated_weight_array.shape[0] / updated_weight_array.shape[1])) # sqrt(Out/In)
+                scale = jnp.sqrt(
+                    jnp.maximum(1, updated_weight_array.shape[0] / updated_weight_array.shape[1])
+                )  # sqrt(Out/In)
             else:
                 scale = 0.2 * jnp.sqrt(jnp.maximum(updated_weight_array.shape[0], updated_weight_array.shape[1]))
             updated_weight_array *= scale
@@ -174,7 +190,7 @@ def zeropower_via_newtonschulz5(X, steps=5, eps=1e-7):
     # really do something even fancier.
     # It would be even smarter to stack similar layers together, but that would require more even more work
     # Let's call this good enough until we think it's not good enough
-    if len(haliax.partitioning._get_mesh().devices):
+    if not jax.sharding.get_abstract_mesh().empty:
         X = jax.lax.with_sharding_constraint(X, PartitionSpec(None, ("data", "model")))
 
     for i in range(steps):
