@@ -1,3 +1,6 @@
+# Copyright 2025 The Levanter Authors
+# SPDX-License-Identifier: Apache-2.0
+
 import contextlib
 import dataclasses
 import logging
@@ -16,6 +19,7 @@ from levanter.tracker import CompositeTracker, Tracker
 from levanter.tracker.helpers import hparams_to_dict
 from levanter.tracker.histogram import Histogram
 from levanter.tracker.tensorboard import TensorboardTracker
+from levanter.tracker.trackio import TrackioTracker
 from levanter.tracker.tracker import DictTracker
 from levanter.tracker.wandb import WandbTracker
 from levanter.utils.jax_utils import is_inside_jit
@@ -196,7 +200,7 @@ def log_configuration(hparams: Any, config_name: Optional[str] = None):
                     name = config_name or "config.yaml"
                     _global_tracker.log_artifact(config_path, name=name, type="config")
             except Exception:  # noqa
-                logger.warning("Failed to dump config to yaml. Skipping logging as artifact.")
+                logger.warning("Failed to dump config to yaml. Skipping logging as artifact.", exc_info=True)
 
 
 def set_global_tracker(tracker: Tracker):
@@ -224,8 +228,7 @@ def set_global_tracker(tracker: Tracker):
 
 
 @typing.overload
-def current_tracker() -> "Tracker":
-    ...
+def current_tracker() -> "Tracker": ...
 
 
 @typing.overload
@@ -265,18 +268,19 @@ def current_tracker(
 
 
 @typing.overload
-def get_tracker(name: Literal["wandb"]) -> WandbTracker:
-    ...
+def get_tracker(name: Literal["wandb"]) -> WandbTracker: ...
 
 
 @typing.overload
-def get_tracker(name: Literal["tensorboard"]) -> TensorboardTracker:
-    ...
+def get_tracker(name: Literal["tensorboard"]) -> TensorboardTracker: ...
 
 
 @typing.overload
-def get_tracker(name: str) -> Tracker:
-    ...
+def get_tracker(name: Literal["trackio"]) -> TrackioTracker: ...
+
+
+@typing.overload
+def get_tracker(name: str) -> Tracker: ...
 
 
 def get_tracker(name: str) -> Tracker:
