@@ -1,3 +1,6 @@
+# Copyright 2025 The Levanter Authors
+# SPDX-License-Identifier: Apache-2.0
+
 import contextlib
 import json
 import logging
@@ -12,8 +15,8 @@ import fsspec
 import jax
 import requests  # type: ignore
 from fsspec import AbstractFileSystem
-from jax.experimental.multihost_utils import sync_global_devices
 
+from levanter.utils.jax_utils import sync_global_devices
 
 logger = logging.getLogger(__name__)
 
