@@ -1,3 +1,6 @@
+# Copyright 2025 The Levanter Authors
+# SPDX-License-Identifier: Apache-2.0
+
 from transformers import AutoTokenizer
 
 from levanter.data.packing import PromptCompletion
@@ -15,22 +18,32 @@ def test_iterate_tokenized_requests_with_chat_template():
     if hf_tokenizer.pad_token is None:
         hf_tokenizer.pad_token = hf_tokenizer.eos_token
 
-    # Create chat-like requests
+    # Create chat-like requests with contexts formatted via the chat template, matching harness usage
+    raw_contexts = [
+        "What's the best way to learn Python?",
+        "Explain quantum computing.",
+    ]
+    completions = [
+        " Practice by building small projects.",
+        " Quantum computing uses quantum bits or qubits.",
+    ]
+
+    formatted_contexts = [
+        hf_tokenizer.apply_chat_template(
+            [{"role": "user", "content": context}], tokenize=False, add_generation_prompt=True
+        )
+        for context in raw_contexts
+    ]
+
     requests = [
         Instance(
             request_type="loglikelihood",
             doc={},
-            arguments=("What's the best way to learn Python?", " Practice by building small projects."),
-            idx=0,
-            metadata=("test_task", 0, None),
-        ),
-        Instance(
-            request_type="loglikelihood",
-            doc={},
-            arguments=("Explain quantum computing.", " Quantum computing uses quantum bits or qubits."),
-            idx=1,
-            metadata=("test_task", 1, None),
-        ),
+            arguments=(formatted_contexts[i], completions[i]),
+            idx=i,
+            metadata=("test_task", i, None),
+        )
+        for i in range(len(raw_contexts))
     ]
 
     # Parameters
@@ -38,7 +51,7 @@ def test_iterate_tokenized_requests_with_chat_template():
     batch_size = 2
 
     # Run with chat template
-    results = list(_iterate_tokenized_requests(requests, hf_tokenizer, max_len, batch_size, apply_chat_template=True))
+    results = list(_iterate_tokenized_requests(requests, hf_tokenizer, max_len, batch_size))
 
     # Verify the results
     assert len(results) == len(requests)
@@ -67,16 +80,15 @@ def test_iterate_tokenized_requests_with_chat_template():
         prompt_tokens = result.ids[: result.prompt_length]
         prompt_decoded = hf_tokenizer.decode(prompt_tokens)
 
-        # The prompt should contain key parts of the chat context
-        # For a chat template, this typically includes formatting like "<s>[INST]"
-        if "<|start_header_id|>" in expected_chat_context:
+        # The prompt should contain key parts of the chat context generated by the template
+        if "<|start_header_id|>" in context:
             assert (
                 "<|start_header_id|>" in prompt_decoded
             ), f"Chat template formatting not found in prompt for example {i}"
 
         # Alternative check: the prompt should be longer than the original context
         # due to the chat template adding formatting
-        original_context_tokens = hf_tokenizer(context, truncation=False, padding=False)["input_ids"]
+        original_context_tokens = hf_tokenizer(raw_contexts[i], truncation=False, padding=False)["input_ids"]
         assert result.prompt_length > len(
             original_context_tokens
         ), f"Prompt length not increased by chat template for example {i}"
