@@ -1,3 +1,6 @@
+# Copyright 2025 The Levanter Authors
+# SPDX-License-Identifier: Apache-2.0
+
 import os
 import tempfile
 
@@ -9,7 +12,7 @@ import haliax
 import levanter.main.eval_lm as eval_lm
 import tiny_test_corpus
 from levanter.checkpoint import save_checkpoint
-from levanter.distributed import RayConfig
+from levanter.distributed import DistributedConfig, RayConfig
 from levanter.models.llama import LlamaConfig, LlamaLMHeadModel
 from levanter.tracker import NoopConfig
 from levanter.trainer_state import TrainerState
@@ -47,6 +50,7 @@ def test_eval_lm():
                     max_eval_batches=1,
                     tracker=NoopConfig(),
                     require_accelerator=False,
+                    distributed=DistributedConfig(initialize_jax_distributed=False),
                     ray=RayConfig(auto_start_cluster=False),
                 ),
                 checkpoint_path=f"{f}/ckpt",
@@ -91,6 +95,7 @@ def test_eval_lm_from_hf():
                     max_eval_batches=1,
                     tracker=NoopConfig(),
                     require_accelerator=False,
+                    distributed=DistributedConfig(initialize_jax_distributed=False),
                     ray=RayConfig(auto_start_cluster=False),
                 ),
                 checkpoint_path=f"{f}/ckpt",
