@@ -1,7 +1,12 @@
+# Copyright 2025 The Levanter Authors
+# SPDX-License-Identifier: Apache-2.0
+
 import os
 import tempfile
+import uuid
 
 import equinox as eqx
+import fsspec
 import jax
 import jax.numpy as jnp
 import jmp
@@ -11,11 +16,17 @@ from chex import assert_trees_all_close, assert_trees_all_equal
 from jax.random import PRNGKey
 
 from haliax import Axis
-from haliax.state_dict import ModuleWithStateDictSerialization
+from haliax.state_dict import ModuleWithStateDictSerialization, to_torch_compatible_state_dict
 
-from levanter.compat.hf_checkpoints import SAFE_TENSORS_MODEL, ModelWithHfSerializationMixin, _convert_to_jnp
+from levanter.compat.hf_checkpoints import (
+    SAFE_TENSORS_INDEX_NAME,
+    SAFE_TENSORS_MODEL,
+    ModelWithHfSerializationMixin,
+    _convert_to_jnp,
+)
 from levanter.models.gpt2 import Gpt2Config, Gpt2LMHeadModel
 from test_utils import skip_if_no_torch
+from tests.test_utils import use_test_mesh
 
 
 @skip_if_no_torch
@@ -42,16 +53,18 @@ def test_save_sharded_checkpoints():
     nano_model = mp.cast_to_param(nano_model)
 
     with tempfile.TemporaryDirectory() as tmpdir:
-        converter.save_pretrained(nano_model, tmpdir, max_shard_size=1024)
+        with use_test_mesh():
+            converter.save_pretrained(nano_model, tmpdir, max_shard_size=1024)
 
         # make sure we saved a few different files
         import glob
 
         assert len(glob.glob(tmpdir + "/*.safetensors")) > 1
 
-        loaded_model = converter.load_pretrained(
-            Gpt2LMHeadModel, ref=tmpdir, config=nano_model.config, dtype=mp.param_dtype
-        )
+        with use_test_mesh():
+            loaded_model = converter.load_pretrained(
+                Gpt2LMHeadModel, ref=tmpdir, config=nano_model.config, dtype=mp.param_dtype
+            )
 
         assert loaded_model.config == nano_model.config
         assert loaded_model.Vocab == nano_model.Vocab
@@ -116,46 +129,17 @@ def test_save_pretrained_with_custom_dtype():
     assert wrapped_model.a_bool_buffer.dtype == jnp.bool_
 
     with tempfile.TemporaryDirectory() as tmpdir:
-        # Save the wrapped_model with dtype=jnp.bfloat16
-        # Note: converter.save_pretrained expects a ModelWithHfSerializationMixin,
-        # so we pass the gpt_model, but we want its state_dict to be part of the larger
-        # state_dict from wrapped_model. This requires a bit of manual handling for the test.
-        # The actual dtype conversion logic in _save_pretrained_local uses to_torch_compatible_state_dict,
-        # which will pick up all params if called on wrapped_model.
-
-        # Let's simulate how HFCheckpointConverter would get the state_dict from the *actual* model being saved (gpt_model)
-        # but then for the test, we want to create a combined state dict.
-        # This is a bit of a hack for the test setup because HFCheckpointConverter is designed for LmWithHfSerializationMixin.
-        # The core logic we are testing (_save_pretrained_local's dtype conversion) works on a generic state_dict.
-
-        # For a cleaner test of _save_pretrained_local's specific behavior, we could directly call it
-        # with a manually constructed state_dict.
-        # However, to test the full converter.save_pretrained path, we need a ModelWithHfSerializationMixin.
-
-        # Let's adjust: we'll save the gpt_model, but then inspect a state_dict that *would have been*
-        # created if TestModelWrapper were the one being serialized by to_torch_compatible_state_dict.
-        # This is getting a bit convoluted.
-
-        # Simpler approach: The dtype conversion happens in _save_pretrained_local on whatever state_dict it receives.
-        # HFCheckpointConverter.save_pretrained calls to_torch_compatible_state_dict(model_to_save).
-        # So, if we want to test the selective conversion, the model_to_save (gpt_model here)
-        # itself should contain these diverse dtypes, or to_torch_compatible_state_dict should be
-        # patched/mocked for this test to return a dict with diverse types.
-
-        # Let's use the TestModelWrapper and directly call _save_pretrained_local,
-        # as this directly tests the target logic with a controlled state_dict.
-        # We'll need to save a dummy config.json as _save_pretrained_local expects it.
         os.makedirs(tmpdir, exist_ok=True)
 
-        # This is what we're testing the internals of:
-        converter._save_pretrained_local(
-            wrapped_model,
-            tmpdir,
-            save_tokenizer=False,
-            save_reference_code=False,
-            max_shard_size=10e9,
-            dtype=jnp.bfloat16,
-        )
+        with use_test_mesh():
+            converter.save_pretrained(
+                wrapped_model,
+                tmpdir,
+                save_tokenizer=False,
+                save_reference_code=False,
+                max_shard_size=int(10e9),
+                dtype=jnp.bfloat16,
+            )
 
         saved_file = os.path.join(tmpdir, SAFE_TENSORS_MODEL)
         assert os.path.exists(saved_file)
@@ -172,7 +156,6 @@ def test_save_pretrained_with_custom_dtype():
         assert tensors.get_tensor("an_int_param").dtype == jnp.int32
         assert tensors.get_tensor("a_bool_buffer").dtype == jnp.bool_
 
-        # (Optional) Attempt to load the model back using the converter
         # This part is tricky because load_pretrained is for LmWithHfSerializationMixin and expects a certain structure.
         # We saved a TestModelWrapper's state_dict.
         # For now, verifying the saved file's dtypes is the primary goal for this unit test.
@@ -205,9 +188,14 @@ def test_save_pretrained_default_dtype():
         # Save the wrapped_model without passing dtype
         # Similar to the above test, using _save_pretrained_local for direct testing.
 
-        converter._save_pretrained_local(
-            wrapped_model, tmpdir, save_tokenizer=False, save_reference_code=False, max_shard_size=10e9, dtype=None
-        )  # No dtype override
+        with use_test_mesh():
+            converter.save_pretrained(
+                wrapped_model,
+                tmpdir,
+                save_tokenizer=False,
+                save_reference_code=False,
+                max_shard_size=int(10e9),
+            )  # No dtype override
 
         saved_file = os.path.join(tmpdir, SAFE_TENSORS_MODEL)
         assert os.path.exists(saved_file)
@@ -219,3 +207,55 @@ def test_save_pretrained_default_dtype():
         assert tensors.get_tensor("a_float_param").dtype == expected_float_dtype
         assert tensors.get_tensor("an_int_param").dtype == jnp.int32
         assert tensors.get_tensor("a_bool_buffer").dtype == jnp.bool_
+
+
+def test_save_pretrained_to_memory_fs():
+    fs = fsspec.filesystem("memory")
+    path = f"memory://levanter/hf-save/{uuid.uuid4().hex}"
+
+    gpt2_config = Gpt2Config(num_layers=4, num_heads=1, hidden_dim=32, use_flash_attention=False)
+    converter = gpt2_config.hf_checkpoint_converter()
+    model = Gpt2LMHeadModel.init(converter.Vocab, gpt2_config, key=PRNGKey(4))
+
+    try:
+        fs.rm(path, recursive=True)
+    except FileNotFoundError:
+        pass
+
+    with use_test_mesh():
+        converter.save_pretrained(
+            model,
+            path,
+            max_shard_size=128,
+            save_tokenizer=False,
+            save_reference_code=False,
+            save_feature_extractor=False,
+        )
+
+    stored_files = {fs._strip_protocol(file) for file in fs.find(path)}
+    base_path = fs._strip_protocol(path)
+    safetensor_files = {file for file in stored_files if file.endswith(".safetensors")}
+
+    assert len(safetensor_files) > 1
+    assert f"{base_path}/config.json" in stored_files
+    assert f"{base_path}/{SAFE_TENSORS_INDEX_NAME}" in stored_files
+
+    for file in safetensor_files:
+        with fs.open(file, "rb") as fh:
+            assert fh.read(1) != b""
+
+    with tempfile.TemporaryDirectory() as tmpdir:
+        local_path = os.path.join(tmpdir, "model")
+        fs.get(f"{base_path}/", local_path, recursive=True)
+
+        with use_test_mesh():
+            reloaded_model = converter.load_pretrained(Gpt2LMHeadModel, ref=local_path, config=gpt2_config)
+
+        original_state = to_torch_compatible_state_dict(model)
+        reloaded_state = to_torch_compatible_state_dict(reloaded_model)
+        assert original_state.keys() == reloaded_state.keys()
+        for key, original_value in original_state.items():
+            reloaded_value = reloaded_state[key]
+            assert_trees_all_close(original_value, reloaded_value)
+
+    fs.rm(base_path, recursive=True)
