@@ -1,25 +1,24 @@
+# Copyright 2025 The Levanter Authors
+# SPDX-License-Identifier: Apache-2.0
+
 import functools
 
 import equinox
 import jax
-import numpy as np
 from jax.random import PRNGKey
-from jax.sharding import Mesh
-
 import haliax as hax
 from haliax.partitioning import ResourceAxis
 
 import levanter.tracker.histogram
-from test_utils import skip_if_not_enough_devices
+from test_utils import use_test_mesh, skip_if_not_enough_devices
 
 
 def test_sharded_histogram_simple():
-    mesh = Mesh((jax.devices()), (ResourceAxis.DATA,))
 
     Batch = hax.Axis("batch", 64)
     Feature = hax.Axis("feature", 128)
 
-    with mesh, hax.axis_mapping({"batch": ResourceAxis.DATA}):
+    with use_test_mesh(), hax.axis_mapping({"batch": ResourceAxis.DATA}):
         a = hax.random.normal(PRNGKey(1), (Batch, Feature))
         a = hax.shard(a)
         hist, bins = levanter.tracker.histogram.sharded_histogram(a, bins=32)
@@ -32,12 +31,13 @@ def test_sharded_histogram_simple():
 
 @skip_if_not_enough_devices(2)
 def test_sharded_histogram_tp():
-    mesh = Mesh(np.array(jax.devices()).reshape(-1, 2), (ResourceAxis.DATA, ResourceAxis.MODEL))
-
     Batch = hax.Axis("batch", 64)
     Feature = hax.Axis("feature", 128)
 
-    with mesh, hax.axis_mapping({"batch": ResourceAxis.DATA, "feature": ResourceAxis.MODEL}):
+    with (
+        use_test_mesh(tensor_parallelism=2),
+        hax.axis_mapping({"batch": ResourceAxis.DATA, "feature": ResourceAxis.MODEL}),
+    ):
         a = hax.random.normal(PRNGKey(0), (Batch, Feature)) * 100
         a = hax.shard(a)
         hist, bins = levanter.tracker.histogram.sharded_histogram(a, bins=64)
@@ -49,8 +49,6 @@ def test_sharded_histogram_tp():
 
 
 def test_sharded_histogram_with_vmap():
-    mesh = Mesh((jax.devices()), (ResourceAxis.DATA,))
-
     Layer = hax.Axis("layer", 4)
     Batch = hax.Axis("batch", 16)
     Feature = hax.Axis("feature", 128)
@@ -64,7 +62,7 @@ def test_sharded_histogram_with_vmap():
         hist, bins = hax.vmap(levanter.tracker.histogram.sharded_histogram, Layer)(a, bins=32)
         return hist, bins
 
-    with mesh, hax.axis_mapping({"batch": ResourceAxis.DATA}):
+    with use_test_mesh(), hax.axis_mapping({"batch": ResourceAxis.DATA}):
         a = hax.random.normal(PRNGKey(1), (Layer, Batch, Feature))
         a = hax.shard(a)
         hist, bins = jit_vmap_hist(a)
