@@ -1,3 +1,6 @@
+# Copyright 2025 The Levanter Authors
+# SPDX-License-Identifier: Apache-2.0
+
 import tempfile
 
 import chex
@@ -8,6 +11,8 @@ import pytest
 from jax import numpy as jnp
 
 import haliax
+
+from test_utils import use_test_mesh
 
 from levanter.optim import AdamConfig
 from levanter.optim.skipstep import SkipStepConfig, SkipStepState
@@ -380,7 +385,7 @@ def test_skip_step_state_serialization_can_load_non_skip():
     wrapped_optimizer = skip_conf.wrap(optimizer)
     wrapped_state = wrapped_optimizer.init(model)
 
-    with tempfile.TemporaryDirectory() as tmpdir, jax.sharding.Mesh(jax.devices(), ("device",)):
+    with tempfile.TemporaryDirectory() as tmpdir, use_test_mesh():
         tree_serialize_leaves_tensorstore(tmpdir, initial_state)
 
         restored_state = tree_deserialize_leaves_tensorstore(tmpdir, initial_state)
